Re-factoring to allow extra UBRN matching rules.  WORK IN PROGRESS.
